Fixed everything red

+++ zkwebui/WEB-INF/src/org/adempiere/webui/component/NumberBox.java
@@ -134,7 +134,7 @@ public class NumberBox extends Div
     public void setValue(Object value)
     {
     	if (value == null)
-    		decimalBox.setValue(null);
+    		decimalBox.setValue((BigDecimal) null);
     	else if (value instanceof BigDecimal)
     		decimalBox.setValue((BigDecimal) value);
     	else if (value instanceof Number)
